✅ 자바스크립트로 작성된 테스트 코드 타입스크립트로 변경
- tsconfig exclude에 test 디렉토리 제거
