+++ __tests__/ApplicationTest.test.ts
@@ -1,0 +1,62 @@
+import App from '../src/App';
+import { Console, Random } from '@woowacourse/mission-utils';
+
+const mockQuestions = (answers: string[]) => {
+  Console.readLine = jest.fn();
+  answers.reduce((acc, input) => {
+    return acc.mockImplementationOnce((_, callback) => {
+      callback(input);
+    });
+  }, Console.readLine);
+};
+
+const mockRandoms = (numbers: number[]) => {
+  Random.pickNumberInRange = jest.fn();
+  numbers.reduce((acc, number) => {
+    return acc.mockReturnValueOnce(number);
+  }, Random.pickNumberInRange);
+};
+
+const getLogSpy = () => {
+  const logSpy = jest.spyOn(Console, 'print');
+  logSpy.mockClear();
+  return logSpy;
+};
+
+describe('숫자 야구 게임', () => {
+  test('게임 종료 후 재시작', () => {
+    const randoms = [1, 3, 5, 5, 8, 9];
+    const answers = ['246', '135', '1', '597', '589', '2'];
+    const logSpy = getLogSpy();
+    const messages = [
+      '낫싱',
+      '3스트라이크',
+      '1볼 1스트라이크',
+      '3스트라이크',
+      '게임 종료',
+    ];
+
+    mockRandoms(randoms);
+    mockQuestions(answers);
+
+    const app = new App();
+    app.play();
+
+    messages.forEach((output) => {
+      expect(logSpy).toHaveBeenCalledWith(expect.stringContaining(output));
+    });
+  });
+
+  test('예외 테스트', () => {
+    const randoms = [1, 3, 5];
+    const answers = ['1234'];
+
+    mockRandoms(randoms);
+    mockQuestions(answers);
+
+    expect(() => {
+      const app = new App();
+      app.play();
+    }).toThrow();
+  });
+});
