+++ tsconfig.json
@@ -34,5 +34,5 @@
     "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
   "include": ["src"],
-  "exclude": ["node_modules", "__test__"]
+  "exclude": ["node_modules"]
 }
